Moved platform common code to the PlatformHelper interface
