+++ fabric/src/main/java/org/mtr/mod/render/RenderVehicleHelper.java
@@ -12,8 +12,7 @@ import org.mtr.mapping.mapper.GraphicsHolder;
 import org.mtr.mod.Init;
 import org.mtr.mod.Items;
 import org.mtr.mod.block.BlockPSDAPGDoorBase;
-import org.mtr.mod.block.BlockPlatform;
-import org.mtr.mod.block.BlockPlatformSlab;
+import org.mtr.mod.block.PlatformHelper;
 import org.mtr.mod.client.VehicleRidingMovement;
 
 import javax.annotation.Nullable;
@@ -54,7 +53,7 @@ public class RenderVehicleHelper {
 					final BlockPos checkPos = Init.newBlockPos(checkX, checkY, checkZ);
 					final BlockState blockState = clientWorld.getBlockState(checkPos);
 					final Block block = blockState.getBlock();
-					if (block.data instanceof BlockPlatform || block.data instanceof BlockPlatformSlab) {
+					if (block.data instanceof PlatformHelper) {
 						canOpenDoors = true;
 					} else if (block.data instanceof BlockPSDAPGDoorBase && blockState.get(new Property<>(BlockPSDAPGDoorBase.UNLOCKED.data))) {
 						canOpenDoors = true;
