+++ fabric/src/main/resources/assets/mtr/lang/en_us.json
@@ -88,13 +88,13 @@
 	"block.mtr.platform_indented": "Platform (Hong Kong)",
 	"block.mtr.platform_na_1": "Platform (North America)",
 	"block.mtr.platform_na_1_indented": "Platform (North America)",
+	"block.mtr.platform_na_1_slab": "Platform Slab (North America)",
 	"block.mtr.platform_na_2": "Platform (North America)",
 	"block.mtr.platform_na_2_indented": "Platform (North America)",
+	"block.mtr.platform_na_2_slab": "Platform Slab (North America)",
+	"block.mtr.platform_slab": "Platform Slab (Hong Kong)",
 	"block.mtr.platform_uk_1": "Platform (United Kingdom)",
 	"block.mtr.platform_uk_1_indented": "Platform (United Kingdom)",
-	"block.mtr.platform_slab": "Platform Slab (Hong Kong)",
-	"block.mtr.platform_na_1_slab": "Platform Slab (North America)",
-	"block.mtr.platform_na_2_slab": "Platform Slab (North America)",
 	"block.mtr.platform_uk_1_slab": "Platform Slab (United Kingdom)",
 	"block.mtr.psd_door": "Platform Screen Doors",
 	"block.mtr.psd_door_2": "Platform Screen Doors",
