+++ fabric/src/main/resources/data/minecraft/tags/blocks/mineable/pickaxe.json
@@ -49,15 +49,15 @@
 		"mtr:pids_pole",
 		"mtr:platform",
 		"mtr:platform_indented",
+		"mtr:platform_slab",
 		"mtr:platform_na_1",
 		"mtr:platform_na_1_indented",
+		"mtr:platform_na_1_slab",
 		"mtr:platform_na_2",
 		"mtr:platform_na_2_indented",
+		"mtr:platform_na_2_slab",
 		"mtr:platform_uk_1",
 		"mtr:platform_uk_1_indented",
-		"mtr:platform_slab",
-		"mtr:platform_na_1_slab",
-		"mtr:platform_na_2_slab",
 		"mtr:platform_uk_1_slab",
 		"mtr:psd_door",
 		"mtr:psd_glass",
