+++ fabric/src/main/resources/data/mtr/recipes/platform_hk_1_slab.json
