+++ fabric/src/main/resources/data/mtr/recipes/platform_na_1_slab.json
