+++ fabric/src/main/resources/data/mtr/recipes/platform_na_2_slab.json
