+++ fabric/src/main/resources/data/mtr/recipes/platform_uk_1_slab.json
